Addressed comments and modified the TC name
